requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
openai
jinja2
sqlalchemy
psycopg[binary]
alembic
aiosqlite
asyncpg
databases